Cellular fix to context activation, SARA-R422 corner case: force it. (#1041)
A case has turned up where, for SARA-R422, even though the module indicates that it is context activated on the network, i.e. AT+CGACT? returns 1,1, the module's internal clients (e.g. sockets, MQTT and HTTP) do not believe that is so.  What fixes it is to always do an AT+CGACT=1,1 first, i.e. to force context activation.  No idea why this should be necessary; the existing code has been in use for SARA_R422 for several years.

Anyway, this is now done for SARA-R422.