ESP-IDF change only: proper minimal sdkconfig.defaults for runner build. (#1045)
The sdkconfig.defaults file used with the runner build employed during testing used to contain many, many CONFIG_BLAH fields. This was due to a misunderstanding: I had thought that all CONFIG_BLAH values that might be required in a build had to be present in sdkconfig.defaults (since that's what ESP-IDF saves when you create a sdkconfig.defaults file), I hadn't realised that it was processed, combined, re-combined, wrapped in palm fronds and then lightly toasted before ending up in the build, so actually only those values that are _different_ from the default defaults need to be included.

The sdkconfig.defaults file used with the runner build now only specifies the things that ubxlib cares about.