AT Client and cellular change: allow more timings to be set, reduce some. (#1046)
The AT Client gets new functions to get/set the timeout employed during URC handling and to get/set the "no stutter" retry delay when reading from a UART.

For cellular, these, plus the AT timeout and the AT command send delay, are exposed through new functions uCellAtCommandDelaySet(), uCellAtCommandDelayGet() and uCellAtCommandDelaySetDefault().  uCellAtCommandDelaySet() should only be used WITH EXTREME CAUTION, if you know what you're doing, see the health warning above it.  The uCellAtCommandDelaySet()/uCellAtCommandDelayGet() functions introduced in commit a12674d05317377550686f34af25444f832fc172 are now deprecated and may be removed.

In addition, the delay between AT commands for SARA-R422 and LARA-R6 has been reduced from 100 ms to 20 ms; the 100 ms value was historically applied in line with the SARA-R41x series modules, which definitely needed it, but testing has suggested that 20 ms (which is the value recommended in the AT manual) is fine for SARA-R422 and LARA-R6.