Cellular update: Cellular helper function to set RF bands (#1048)
An easier-to-use interface for the setting of the RF bands employed by a cellular module. The cell configuration gets a new API named uCellCfgSetBands() to simplify cellular RF band setting. This API uses band numbers instead of bitmask. The users can send the desired band numbers to be set in the form of an array to this API.

This API will overwrite the band setting in the cellular module.

Writing numBands to 0 for LENA R8 enable all bands. Some modules don't allow this while others do e.g., LARA R6 allows for disabling all bands.