Zephyr platform change only: update to nrfConnect 2.5.0/Zephyr 3.4.99-ncs.
The nrfConnect/Zephyr version that we test with is updated to 2.5.0/3.4.99-ncs; the only change in the ubxlib code is to remove the now unused cryptographic functions in the port layer (see previous commit).

Note: the minimum version of "west" is now 0.13; if you need to perform an nRFConnect/Zephyr version update yourself and happen to have west 0.12 or earlier, it is worth updating your version of west (using pip) BEFORE you do the installation; we found that installation can fail unless at least version 0.13 of west is _already_ present (since west is used by the installation process itself).