Test automation change only: add --features to invoke. (#1050)
When running invoke from the command line, the switch --features can now be applied to set the feature list, e.g. --features="cell short_range" to leave out gnss.  This overrides the features set by the UBXLIB_FEATURES environment variable.  A "features" field can also be included in the file u_flags.yml in the .vscode directory and this will behave in the same way.