CodeChecker update resulting in LOTS of small changes to keep CodeChecker happy. (#1051)
When the docker image within which the automated tests run was updated earlier this year the new version of CodeChecker that was installed (6.22.2) came with a weird problem: when run on the STM32F4 basis (i.e. test instances 6.2.0 and 6.2.1; the Zephyr 6.1 instance was fine), for every single file it reported:

"ERROR: ld.so: object '/usr/local/lib/python3.8/dist-packages/codechecker_analyzer/ld_logger/lib/x86_64/ldlogger.so' from LD_PRELOAD cannot be preloaded (cannot open shared object file): ignored."

Since no files ended up being parsed the result was zero errors and so this was not noticed.

Turns out this was a problem in that release of CodeChecker and the fix was to install the latest version (6.23) however this, of course, brought with it more checks, hence the rather large number of changes in this commit (nothing of any real significance though) to make it happy.