Test change only: move Wi-Fi Geofence. (#1064)
For whatever reason, the Google location API sometimes gives us a radius of uncertainty of 6 km on the Wifi APs in our office.  Our office isn't that big, and such a huge error messes up the geofence test, since the circles used in it fall inside the area of uncertainty, hence with this commit we move the "outie" circle rather further away.