Test change only: HTTP clean-up. (#1065)
The HTTP Client test, if it hits a failure, wasn't closing any HTTP contexts that were open in the HTTP clean-up "test".  This becomes noticeable if the failure is quite late in the HTTPS part of the test as the security contexts are all used, resulting in subsequent failures in the sockets tests 'cos they are unable to obtain a security context.

A better clean-up is now done.