BREAKING CHANGE to uNetwork API for cellular only: add pMccMnc. (#1063)
The parameter pMccMnc is added to the [end of the] structure uNetworkCfgCell_t; this may be used to force cellular network selection to a specific network (e.g. "23410", i.e. just as the parameter to uCellNetConnect()/uCellNetRegister() of the same name), rather than automatically selecting the best available network.

Provided your network configuration structure is a const, or is initialised to zero, this will not affect you. OTHERWISE PLEASE ENSURE THAT THIS PARAMETER IS SET TO NULL if you still wish to automatically select the best available network, as before.