Sockets fix: free security context on clean-up. (#1075)
When a secured socket is closed the TLS security context is, correctly, free'd. However, if uSockCleanUp() happened to be called before a socket, that is being asynchronously closed, is finally closed, it would remove all record of the socket, meaning that the TLS security context would not be free'd when the closure finally occurs.

This is fixed by also freeing the TLS security context in uSockCleanUp().

Thanks to FloriVeit/Johannesmay for pointing this out.