Windows platform change only: fix critical section hole. (#1076)
The functions uPortEnterCritical()/uPortExitCritical(), which suspend/resume all but the current task, use a list of tasks maintained by uPortTaskCreate()/uPortTaskDelete() and all of these functions individually lock a single gMutexThread to ensure re-entrancy.

However, very occasionally, the function uPortPrivateEnterCritical() asserts that it is already in a critical section, even though it cannot be, pointing to a flaw in the logic.

The flaw is that it is possible for uPortTaskCreate() or uPortTaskDelete() to be called DURING a critical section, i.e. after uPortEnterCritical() has returned and before uPortExitCritical() has been called; a task created at such a time will NOT be suspended, effectively breaking the critical section; not sure what deleting a task during that time would do but it is probably not good.

To fix this, uPortEnterCritical() now locks gMutexThread but does NOT unlock it, uPortExitCritical() performing the unlock part; this fixes the hole.