Add support for u-Connect 2nd generation (#1079)
This commit adds functionality for the 2nd generation of
u-Connect for short range modules. Some refactoring
of the directory structure has been made. All short
range source files which contain AT commands have
been given duplicates to be used when the new u-Connect
is to be used. These files have the same name but are
stored in sub directories named "gen2". To handle this
a new ubxlib feature named "short_range_gen2" has been
added. When present the build will automatically use
the gen2 files instead.

A new git submodule which encapsulates much of the
AT-command handling is used in this case.

The only available module for this right now is Nora W36.
The current firmware for this has some limitations which
means that some tests have been disabled for this module.