Cellular API addition: SIM FPLMN list delete. (#1089)
This commit introduces a [minimal] SIM API to the cellular API, specifically to allow the FPLMN (Forbidden Public Land Mobile Network) list on the SIM to be deleted. This SIM API is deliberately minimal, since the kinds of application that employ ubxlib do not use a SIM PIN (this offers little security since it has to be stored on an MCU and sent across a UART interface when the module is powered-up), don't require phone-book access, etc.

If you would like more SIM-related features, please let us know.