HTTP fixes for cellular: timeout cannot be zero. (#1086)
Cellular modules return an error if an attempt is made to set the HTTP timeout to zero:

AT+UHTTP=0,7,0
+CME ERROR: Operation not supported

This is pointed out in the description of uCellHttpOpen() and does not occur if the intended pattern of assigning uHttpClientConnection_t to U_HTTP_CLIENT_CONNECTION_DEFAULT when calling pUHttpClientOpen() is followed. However that pattern is not documented in the description of pUHttpClientOpen() and uCellHttpOpen() allows the timeout to be set to zero, which it should not.

Our thanks to jamal-mouline for highlighting this.