POSSIBLY BREAKING CHANGE FOR ZEPHYR: more UARTs. (#1100)
In order to support non-Nordic MCUs (e.g. STM32 MCUs) that have more than four UARTs, the number of UARTs supported by the UART code is increased to eight. In order not to waste RAM, the array that was previously used to store the UART data is changed to be a linked list.

THE POSSIBLY BREAKING PART: in making this change we modified the static interrupt callback function, uartCb(), to receive a pointer to the UART data as its user parameter, rather than having to search through the array (now the linked list) for the correct entry; this seemed like a sensible efficiency improvement.  HOWEVER, with this done, UART transfers started failing; turns out that the delay introduced by the look-up code was serving a purpose: without it the interrupt does not work correctly.  Hence there is now a wait of U_ZEPHYR_PORT_UART_INTERRUPT_DRIVEN_NUM_NOPS (50) nops() at the start of uartCb().  This works fine on all of the platforms we test with but IF YOU HAVE YOUR OWN ZEPHYR PORT you may find that you need to increase U_ZEPHYR_PORT_UART_INTERRUPT_DRIVEN_NUM_NOPS (or, of course, you may be able to reduce it).