Zephyr change only: expand range/handling of GPIOs. (#1101)
In a similar vein to commit 259895276aa88ef963cf16c347d83ff8aeb981ab, in order to support non-Nordic (e.g. STM32) MCUs, this commit adds support to the Zephyr platform for up to 10 GPIO ports, with all of the naming options usually employed by the device tree, and correctly calculates the number of pins per port based on the contents of the device tree (e.g. 32 pins per port for nRF52/nRF53, 16 for STM32).