Zephyr platform change only: add native Zephyr STM32 test instance. (#1104)
This commit consists mostly of changes to the ubxlib test system to support native Zephyr, i.e. Zephyr used directly rather than via the nRF Connect SDK. In this particular case Zephyr with the STM32 HAL is tested with a Nucleo F767ZI board (carrying an STM32F7 device). As well as testing non-nRF52/nRF53 devices under Zephyr, this tests configuring ubxlib through the Zephyr device tree on an end-to-end basis, and also tests use of device tree aliases, which handle the fact that ST label some of their UARTs USARTs. Note that the test is currently just of the port layer alone, no devices are [yet] connected to the board, those will be added in subsequent PRs.

The testing has resulted in one very minor improvement in the Zephyr device-tree-based configuration code, which is included here: the cellular network .dts node (i.e. u-blox,ubxlib-network-cellular) can now be omitted, if the application has nothing to say about it.