IMPORTANT ZEPHYR change: fix for priority of SYS_INIT(). (#1106)
ubxlib_preinit(), which is called by the Zephyr SYS_INIT() macro to create the ubxlib thread stacks with a call to k_thread_system_pool_assign(k_current_get()), has always been set to run at boot stage PRE_KERNEL1; this was a requirement of Zephyr when the change was introduced two years ago.

However, it appears that things have moved on, and now the Zephyr documentation states that, at PRE_KERNEL1, "devices cannot use any kernel services during configuration, since the kernel services are not yet available" and, at least on an nRF9160 based board, the return value of k_current_get() seems not to be valid at PRE_KERNEL1.  Changing the boot stage at which ubxlib_preinit() is called to POST_KERNEL fixes this problem.

Our thanks to Irockasingranite for finding this issue and proposing the fix.