Cellular change only: minor fixes for SARA-U201. (#1109)
SARA-U201 is no longer sold by u-blox, however there are customers still using it and its descendants (e.g. SARA-U260). It was not tested in the ubxlib test farm for a few years because the local cellular networks would reject a device that connected/disconnected frequently (something we do during testing); however, we now have SIM cards which work on local networks that do not throw a strop, and hence SARA-U201 can be tested once more.

This commit contains a few tweaks to the core code:

- move an AT timeout setting that was in the wrong place, around the sending of AT+COPS=0; it just happened to work with all other modules, not so SARA-U201,
- the "I've found no networks" response to AT+COPS=?, which SARA-U201 sometimes returns (the other modules tend to return CME ERROR: Temporary failure), is 13 characters long, not 12,
- there was a hole in the logic of uCellCfgSetRatRank(), specific to SARA-U201, where it would not coalesce the RATs back to single RAT mode when duplicate RATs appeared,
- CMUX and PPP are now enabled for SARA-U201, since they are being tested.

There are also a few tweaks related to testing:

- add the ability to not run the security tests, since they use EC encryption which SARA-U201 does not support,
- allow the default APN, used when pApn is set to NULL, to be overridden by the test system to something other than "internet" in order that the examples (which set the APN to NULL) can be made to work on a live network when run inside the test system: LTE networks just seem to sort the APN out, whatever the user requests; not 2G or 3G networks though.