POSSIBLY BREAKING CHANGE for cellular TLS: check number of cipher suites. (#1111)
The ability for the application to select which cipher suites are proposed for a TLS connection is a feature that depends upon the cellular module being used. As is already mentioned in the cellular API, SARA-R5 supports a list of user-selected cipher suites while the other module types do not. However, when a uSecurityTlsSettings_t structure is received, the internal uSecurityTls-level API does not know the capabilities of the underlying module and so it cannot check; therefore an application might select, say, five cipher suites and receive a success response, when in fact only the last of the five cipher suites is actually being used in the connection by the cellular module.

To make the operation of the uSecurityTls API clearer, a function uCellSecTlsCipherSuiteMoreThanOne() is added in the uCellSecTls API; when a uSecurityTlsSettings_t structure is processed by the uSecurityTls API, this function is used to check that the caller has not tried to select more than one cipher suite when that operation is not supported.

THE POSSIBLY BREAKING PART is that an application may be "succeeding" in setting more than one cipher suite when, in fact, what is actually happening is that it is only setting the very last of N and that just happens to work with the TLS stack of the target server. The application will now get back a U_ERROR_COMMON_TOO_BIG error, instead of zero for success.  To return to a success outcome, preserving the old behaviour (assuming that behaviour was and is acceptable) simply delete all bar the last cipher suite in your list and set the number of cipher suites to 1.