Test change only: Docker image and u_packages.yml update. (#1112)
The Docker image is updated to install ppp for Linux, needed for the up-coming addition of support for PPP connectivity via cellular on Linux.  In updating the Docker image it seems that Segger have removed the version of their JLInk tools that we were using (i.e. 7.88n has gone from their website), so we also switch to the latest named version of their tools (7.94m).

Since a new version of PyLint is brought in by this with some new recommendations there are associated Python script updates.