POSSIBLY BREAKING CHANGE for Linux uPortUartPrefix(): range check. (#1113)
Should a prefix be passed to uPortUartPrefix() that was greater than 32 characters in length, it would be accepted and truncated to 32 characters but, potentially, no terminator might be added; the truncation is done with strncpy() but strncpy() doesn't _guarantee_ a terminator, it just guarantees no overrun.  The length is now checked on entry so that a prefix that is too long is rejected; hence a terminator is now guaranteed.  Also the names of a few static functions are updated to follow the pointer syntax used elsewhere.

THE POSSIBLY BREAKING PART is that you may have previously been setting a prefix that was too long and getting away with it: you will now get back U_ERROR_COMMON_INVALID_PARAMETER instead.