Cellular fix for MQTT only: detect SARA-R4-style disconnect. (#1114)
When an MQTT connection is disconnected by the far end, a module using the SARA-R4 old-syntax MQTT AT commands might not send a "logout " URC (i.e. +UUMQTTC: 0,x), it might instead send a "login" URC with failure, i.e. +UUMQTTC: 1,x, where x != 0. This was not being flagged as a disconnect: it is now.