Cellular changes, PPP: enable for LENA-R8, fix ESP-IDF timeout and fix dial-up. (#1116)
PPP support is enabled for LENA-R8, a minor fix is made to the PPP dial-up process and U_PORT_PPP_SHUTDOWN_TIMEOUT_SECONDS is extended to 15 seconds (from 10 seconds) as the time-out within the ESP-IDF LWIP PPP layer is around 10 seconds; if you don't give it enough time to sort itself out in cases where cellular is being brought down without first disconnecting from the network, ESP-IDF will go "bang" somewhere inside LWIP PPP.

IMPORTANT NOTE FOR LENA-R8: on LENA-R8 it is not possible to use the same PDP context for PPP as for AT-command-based operation: if you do so then, once PPP is active, commands such as uSockGetHostByName() and any attempt to use the MQTT or HTTP clients on LENA-R8 will fail.  Hence we set the PDP context for PPP operation to be separate.  It is POSSIBLE that there are cellular networks out there which will not allow more than one PDP context, in which case connections for LENA-R8 will fail when U_CFG_PPP_ENABLE is defined for the build; should this happen then you should compile this code with U_CELL_PRIVATE_PPP_CONTEXT_ID_LENA_R8 set to -1 and the code will use the same PDP context for both (and the on-module clients will not be available to your application while PPP is active).