Cellular change only: updates in cellular any module type power on process (#1110)
The single pulse of 2000 ms for power on the cellular module does not work for all the module types, i.e. some modules (especially SARA-R520) have a behaviour that they are first powered up and then power down again in this time. So, hindering the automatic identification process.
Hence, we have introduced two retries with variable pulse times (defined by `gPwrOnPinDurationMs[]`). By default, they are set to 300ms and 2000ms so as to peacefully power up all the modules supported by ubxlib.
Also a test is added named `uCellPwrAnyModule` for a quick test of the powering on process of "any module type" case.