Cellular MQTT change only: allow AT poke during publish to be disabled. (#1124)
When UART power saving is switched on some modules (e.g. SARA-R422) can sometimes withhold URCs, hence during an MQTT publish we have always poked the module with AT to make sure that it has not gone to sleep on us. Since we are either publishing a message or waiting for an acknowledgement of that publish from the MQTT broker during this time, the module is unlikely to be able to do any sleeping but, if you are especially concerned about power saving, you may now disable the poking by defining U_CELL_MQTT_POKE_DURING_PUBLISH_DISABLE for the build. Of course, you do need to test that you do not have a URCs-held-back issue with your particular module.