Sockets API addition: uSockForgetAll() (#1128)
A function is added, uSockForgetAll(), which will cause this code to forget about all sockets, returning to a state as if none were ever opened. This is useful in a VERY SPECIFIC circumstance: when using sockets with a cellular module and that module has entered 3GPP power saving, losing all RAM contents and hence all sockets. Calling uSockForgetAll() will serve to match the state of this code with that of the module in that situation.