Test change only: swap LARA-R6 from instance 30 to instance 31. (#1131)
The LARA-R6 cellular module on test instance 30 is moved to test instance 31. This is done to make room for SARA-R520 testing on instance 30 where CMUX is known to work. On a temporary basis CMUX testing is disabled on instance 31; turns out that the DMA needed to make UART operation reliable for STM32Fx devices, with their 1-byte UART buffers, requires our Zephyr-platform uPortUart code to be restructure to take advantage of DMA, it is not as simple as just enabling DMA in the device tree.