GNSS debug feature: read messages into uCenter (#1132)
A Python script, u_gnss_create_ucenter_ubx_file.py, is provided which can take the ubxlib log output, find the UBX-format messages in it (the messages sent by the GNSS device) and write these messages in a form that can be read by the u-blox uCenter tool. The input can be a file or can be the live log output from ubxlib, the output can be a file or it can be piped through a pair of virtual COM ports so that it may be opened directly by uCenter itself.