Updated example for cellular only: clearer eDRX outcome. (#1135)
The E-DRX example for cellular expects that, once E-DRX has been set, the module will emit a URC indicating the assigned E-DRX settings. This works internally on the ubxlib test system however it doesn't seem to be a reliable indication for a live network: likely it depends on the network and when the radio connection is released. So as not to confuse customers, the example is now updated to call the uCellPwrGetEDrx() API to determine the truth if the callback happens not to be called.

Thanks to elektron314 for spotting this.