Cellular PPP change only: PPP over different port.
The current implementation of PPP runs over CMUX, ensuring that AT commands and GNSS commands can continue while the PPP link is running.  However, u-blox cellular modules do not support use of CMUX on USB connections; one is intended to use different USB end-points for PPP, AT and GNSS traffic.  This is a problem when using a u-blox module [one that supports AT commands over USB] and only a USB connection is available to the MCU.

With this commit, support is provided for running the PPP interface over a different port to that of the AT interface, which could be a USB end-point, and hence a PPP connection can now be made when only a USB connection to the module is available; see examplePppLinuxSockets() for how to do it.