Linux fix and cosmetic test fix: allow more than one UART set to -1.
The Linux uPortUartOpen() interface allows the UART number to be set to -1, letting the prefix be used as a whole, no number on the end, however the logic wasn't quite right when more than one UART is opened with UART number set to -1; it would think that the second UART had already been opened.  This is now fixed.

A cosmetic fix is also made to the prints in the Zephyr PPP sockets test.