Test change only: disable PPP testing on instance 32 (SARA-U201/Zephyr). (#1140)
As of a few days ago, with no code or HW changes, testing of a TCP socket connection from the Zephyr IP stack, through PPP to SARA-U201 and then on to the TCP echo server at ubxlib.com, (i.e. tests examplePppZephyrSockets() and zephyrSockTcp()) began failing. LCP/IPCP negotiation on the PPP link completes, the echo server sees the TCP connection request (several of them, over a 2 second period) and responds to each one but none of the responses ever come out of SARA-U201 (this checked with a Saleae probe on the UART lines). The same tests run with the IP stack on-board SARA-U201 (i.e. using the sockets AT commands) work fine.

It is difficult to see why this should be: these two tests, with the same code, passed just three days ago, yet the failure is absolutely consistent.

Since SARA-U201 is a deprecated product, tracing it and searching for PPP issues is not a viable way forward, hence U_CFG_PPP_ENABLE is now removed from instance 32. This means that we aren't testing Zephyr PPP; that will be fixed by switching the modules on instances 31 and 32, so LARA-R6 ends up on instance 32, then putting U_CFG_PPP_ENABLE onto the LARA-R6/Zephyr instance. But that is not possible right now as there are several PRs in train that would be broken if the HW were changed underneath them.

Will follow up with that test-only change next week.