IMPORTANT FIX FOR ESP-IDF: correct default max task priority. (#1147)
For the ESP-IDF platform we set U_CFG_OS_PRIORITY_MAX to 25, a value chosen to match the default value of configMAX_PRIORITIES in FreeRTOSConfig.h, which is [still] 25.  However, we were out by one: configMAX_PRIORITIES is the maximum _number_ of priorities, rather than the maximum priority; the maximum priority is configMAX_PRIORITIES - 1, U_CFG_OS_PRIORITY_MAX should be 24.  The ubxlib code has used this value forever and a day without any apparent problem but in ESP-IDF version 5.2 an assert now goes off:

components/freertos/FreeRTOS-Kernel/tasks.c:1061: assert failed: prvInitialiseNewTask tasks.c:1061 (uxPriority < ( 25 ))

If you see this assert, this is the fix.

Our thanks to michaelboeding for finding the problem and testing the fix.