MAJOR CHANGE: uTimeout API. (#1143)
MAJOR CHANGE: uTimeout API.

IN CASE OF COMPILATION FAILURE: if you have copied an existing ubxlib .c file to create your own file, you may find it fails to compile after merging this commit: this is because, in our internal .c files, we do not include ubxlib.h, we bring in only the required headers, and so your .c file will be missing u_timeout.h.  To fix this you should #include "u_timeout.h" nearish the top of the inclusions in your .c file, above where you would include things like u_at_client.h, below where you would include u_error_common.h (see the modified .c files in this commit for the pattern).

A new common API, uTimeout, is added and all timeouts are routed through it. This API performs time comparisons with unsigned 32-bit arithmetic to ensure that timeouts expire correctly if they happen to cross the 32-bit wrap point, which will occur every 50 days for a millisecond tick; otherwise there were instances where a timeout could potentially get "stuck" for 25 days (the unsigned 32-bit wrap length for a millisecond tick).

The implementation of the API includes the ability to speed up the apparent wrap-rate of the underlying tick; this is used on test instance 23 (which includes cellular, GNSS and short-range module types) to give it a good thrashing.

Our thanks to warasilapm for raising this issue and _my_ thanks to mazgch for providing the solution.