Test change only: improved start-up time for u_runner. (#1154)
The u_runner code which pulls together all of the ubxlib tests does so by subtle use of the constructor attribute applied to C code.  This is fine except that constructors are run during C start-up, before main(), and at that time some platforms have not brought their clocks up to full speed; this is the case for STM32 MCUs.  That is also fine but it does mean that one should not do very much work in the constructor; the u_runner application was alphabetically sorting the list of tests each time a new one was added, which with 157 items was beginning to take rather a long time: fully 4 minutes on an STM32U5 processor running at 4 MHz, as it does during startup.

u_runner is now modified to only do the alphabetical sorting when the list of tests is first used.