IMPORTANT CELLULAR FIX for context activation with some MNO profiles. (#1163)
In commit ac08677d4570aa46632317c28a4b114ad645f47d (related to LENA-R8), the code that checks cellular context activation was optimised to reduce the number of times it went around a for() loop.  However, this optimisation did not work for certain MNO profiles, profiles which explicitly require the module to employ a default PDP context (context 0) was well as the requested context, hence the optimisation is removed.

Our thanks to jraats for pointing out the problem.