Zephyr platform fix: allow DT-supplied network config for cellular to be NULL again. (#1164)
It is intended that the network configuration passed to uNetworkInterfaceUp() can be NULL for the case where the network configuration is passed in via the Zephyr device tree.  However, when introducing support for PPP over a different port in commit 16732b7a01fdc8b4fee808f9d3ea64551b5f5a12 we forgot this.  Now fixed.

Our thanks to jraats for spotting this and suggesting the fix.