Documentation change only: describe uNetworkInterfaceDown() corner case.
There is a corner-case for uNetworkInterfaceDown() when it is called on a GNSS network that is inside a cellular device.  It goes like this:

When uNetworkInterfaceUp() is called on a GNSS network, ubxlib tells the cellular device to create a CMUX channel (to carry AT and GNSS traffic simultaneously); the configuration of the existing AT client is copied to the new one and the original AT client is left locked.  If uNetworkInterfaceDown() is called for the GNSS interface on the cellular device from a different task to the one that called uNetworkInterfaceUp(), some RTOSes (e.g. FreeRTOS will do this) raise an assert that a mutex is being unlocked by a task that didn't lock it in the first place, which is true 'cos the AT client lock is, of course, a mutex.

There is nothing, in principle, wrong with an application calling uNetworkInterfaceDown() from different task to that where it called uNetworkInterfaceUp(), so it would be nice to allow this.  However it is also true that leaving the original AT client locked is a good thing, since it prevents it being used accidentally (not sure what would happen if it was but it would probably not be good).

This commit adds a note to the description of uNetworkInterfaceDown() until we can determine a better way forward.

Our thanks to ws998116 for finding this issue and being happy to work around it as described above for now.