Zephyr change only: option to use async Zephyr UART API (e.g. for STM32Fx MCUs). (#1166)
If you are using Zephyr with an STM32Fx-series (e.g. STM32F4 or STM32F7) MCU, these chips have only a single-byte UART receive buffer, which will lose characters occasionally running at 115200; the UART HW blocks in these MCUs are designed assuming DMA will be used and this is how the ubxlib stm32cube platform UART implementation works. However, the normal Zephyr UART driver does not use DMA, only the asynchronous Zephyr UART driver does so.

Hence, with this commit, the option to make the ubxlib UART driver for Zephyr use the asynchronous Zephyr UART API is introduced: to do so, define U_ZEPHYR_PORT_UART_ASYNC (as well as setting CONFIG_UART_ASYNC_API to y in your Zephyr .conf file) when you build ubxlib. It is wise to use this option when running Zephyr on an STM32Fx-series MCU to avoid character loss.

Note that this is not necessary for the STM32U5 series MCUs, which all sport 8-byte HW buffers in the UART-HW-block department.

To be clear: this is NOT a change to the ubxlib uPortUart API, that remains exactly the same, synchronous, as before, it only changes which underlying Zephyr API the uPortUart API implementation for Zephry integrates with.