LENA-R8 fix for GNSS operation: GNSS power-on retries for the "any type" case.
Most cellular modules ensure that any internal GNSS device is powered-up before returning OK to the AT+UPGS command that powers the internal GNSS module up.  However, LENA-R8 does not and, in the case where the ubxlib code has been asked to determine the module type by itself, the GNSS message sent to do that elicits no response and so the whole GNSS power-up process fails.

With this commit, the code retries up to 3 times at 1 second intervals if it gets no response.

Our thanks to mos216 for highlighting this.